requirements.txt
Flask==2.2.3
Werkzeug==2.2.2
gspread==5.4.0
oauth2client==4.1.3
python-dotenv==1.0.0
